WIP revenue line items - need add more and delete functionality
